# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Fikirden Ürüne Proje Vitrini API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new student account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered and logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user session
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List approved projects (public)
      parameters:
        - name: city
          in: query
          schema:
            type: string
            enum: [Konya, Afyonkarahisar]
        - name: category
          in: query
          schema:
            type: string
            enum: [web-tasarim-ve-gelistirme, mobil-uygulama, oyun-gelistirme, yapay-zeka]
        - name: stage
          in: query
          schema:
            type: string
            enum: [MVP, Prototip]
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of approved projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectSummary"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get project detail (public, approved only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/projects:
    get:
      operationId: listAdminProjects
      tags: [admin]
      summary: List all projects for admin review
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected, draft]
      responses:
        "200":
          description: All projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminProject"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/projects/{id}/approve:
    patch:
      operationId: approveProject
      tags: [admin]
      summary: Approve a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "403":
          description: Not admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/projects/{id}/reject:
    patch:
      operationId: rejectProject
      tags: [admin]
      summary: Reject a project with feedback
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectInput"
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "403":
          description: Not admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    AuthUser:
      type: object
      required: [id, email, fullName, school, city, role]
      properties:
        id:
          type: integer
        email:
          type: string
        fullName:
          type: string
        school:
          type: string
        city:
          type: string
          enum: [Konya, Afyonkarahisar]
        role:
          type: string
          enum: [student, admin]
    RegisterInput:
      type: object
      required: [email, password, fullName, school, city]
      properties:
        email:
          type: string
        password:
          type: string
          minLength: 6
        fullName:
          type: string
          minLength: 2
        school:
          type: string
          minLength: 2
        city:
          type: string
          enum: [Konya, Afyonkarahisar]
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
...
    AiUsage:
      type: object
      required: [isUsed, tools, usageAreas, verificationMethod]
      properties:
        isUsed:
          type: boolean
        tools:
          type: array
          items:
            type: string
        usageAreas:
          type: array
          items:
            type: string
        verificationMethod:
          type: string
    SecurityAndEthics:
      type: object
      required: [noRealPersonalData, noSecretsInGithub, copyrightCompliant]
      properties:
        noRealPersonalData:
          type: boolean
        noSecretsInGithub:
          type: boolean
        copyrightCompliant:
          type: boolean
    ProjectSummary:
      type: object
      required: [id, title, slug, category, city, stage, status, coverUrl, techStack, aiUsage, createdAt, studentName]
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        category:
          type: string
          enum: [web-tasarim-ve-gelistirme, mobil-uygulama, oyun-gelistirme, yapay-zeka]
        city:
          type: string
          enum: [Konya, Afyonkarahisar]
        stage:
          type: string
          enum: [MVP, Prototip]
        status:
          type: string
          enum: [pending, approved, rejected, draft]
        coverUrl:
          type: ["string", "null"]
        techStack:
          type: array
          items:
            type: string
        aiUsage:
          $ref: "#/components/schemas/AiUsage"
        createdAt:
          type: string
        studentName:
          type: string
    Project:
      type: object
      required: [id, title, slug, category, city, stage, status, coverUrl, screenshotUrls, techStack, githubUrl, demoUrl, aiUsage, securityAndEthics, createdAt, updatedAt, studentId, studentName, problem, targetAudience, solution, futurePlans, adminFeedback]
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: string
        title:
          type: string
        slug:
          type: string
        category:
          type: string
          enum: [web-tasarim-ve-gelistirme, mobil-uygulama, oyun-gelistirme, yapay-zeka]
        city:
          type: string
          enum: [Konya, Afyonkarahisar]
        stage:
          type: string
          enum: [MVP, Prototip]
        problem:
          type: string
        targetAudience:
          type: string
        solution:
          type: string
        futurePlans:
          type: string
        techStack:
          type: array
          items:
            type: string
        githubUrl:
          type: ["string", "null"]
        demoUrl:
          type: ["string", "null"]
        coverUrl:
          type: ["string", "null"]
        screenshotUrls:
          type: array
          items:
            type: string
        aiUsage:
          $ref: "#/components/schemas/AiUsage"
        securityAndEthics:
          $ref: "#/components/schemas/SecurityAndEthics"
        status:
          type: string
          enum: [pending, approved, rejected, draft]
        adminFeedback:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    AdminProject:
      type: object
      required: [id, title, slug, category, city, stage, status, coverUrl, techStack, aiUsage, createdAt, studentName, studentEmail, adminFeedback]
      properties:
        id:
          type: integer
        studentName:
          type: string
        studentEmail:
          type: string
        title:
          type: string
        slug:
          type: string
        category:
          type: string
          enum: [web-tasarim-ve-gelistirme, mobil-uygulama, oyun-gelistirme, yapay-zeka]
        city:
          type: string
          enum: [Konya, Afyonkarahisar]
        stage:
          type: string
          enum: [MVP, Prototip]
        status:
          type: string
          enum: [pending, approved, rejected, draft]
        coverUrl:
          type: ["string", "null"]
        techStack:
          type: array
          items:
            type: string
        aiUsage:
          $ref: "#/components/schemas/AiUsage"
        adminFeedback:
          type: ["string", "null"]
        createdAt:
          type: string
...
    RejectInput:
      type: object
      required: [feedback]
      properties:
        feedback:
          type: string
          minLength: 5