# lib/api-zod/src/generated/types/adminProject.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Fikirden Ürüne Proje Vitrini API
 * OpenAPI spec version: 0.1.0
 */
import type { AdminProjectCategory } from './adminProjectCategory';
import type { AdminProjectCity } from './adminProjectCity';
import type { AdminProjectStage } from './adminProjectStage';
import type { AdminProjectStatus } from './adminProjectStatus';
import type { AiUsage } from './aiUsage';

export interface AdminProject {
  id: number;
  studentName: string;
  studentEmail: string;
  title: string;
  slug: string;
  category: AdminProjectCategory;
  city: AdminProjectCity;
  stage: AdminProjectStage;
  status: AdminProjectStatus;
  /** @nullable */
  coverUrl: string | null;
  techStack: string[];
  aiUsage: AiUsage;
  /** @nullable */
  adminFeedback: string | null;
  createdAt: string;
}