# lib/api-zod/src/generated/types/securityAndEthics.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Fikirden Ürüne Proje Vitrini API
 * OpenAPI spec version: 0.1.0
 */

export interface SecurityAndEthics {
  noRealPersonalData: boolean;
  noSecretsInGithub: boolean;
  copyrightCompliant: boolean;
}